Update Charging Limiter
Added some comments for better understanding this
Also added a continue value like in apps to let the user choose when charging
should be continued again. Now also bug of endless recharging to stop value is fixed with this
Some exceptions added too, so the user can't set crap values to force bugs.

Later this will be written to an own module (Code is still a bit dirty)