Introduce Charging Controller with night mode (v1)
This is another charging module made by me.
It calculates the needed charging rate by the capacity needed to charge
till 100% or whatever is set for stop charge value of charging limiter
(for now with static voltages -> they
are left out bc the will not be needed in the calculation)

There will be more updates in future.
Now there is only one bug... After setting low charging rates and turning off
night mode, the charging rate will stay low
This is easily fixable with a reboot, but anyways it's an annoying issue
and will be fixed in some of the next builds.